+++ entropy/urandom/Raaz/Entropy.hs
@@ -1,10 +1,16 @@
 -- | Entropy based on `/dev/urandom`.
-module Raaz.Entropy( getEntropy ) where
+module Raaz.Entropy( getEntropy, entropySource ) where
 
 import Control.Monad.IO.Class( MonadIO, liftIO)
 import System.IO
 import Raaz.Core
 
+
+
+-- | The name of the source from which entropy is gathered. For
+-- information purposes only.
+entropySource :: String
+entropySource = "/dev/urandom(generic posix)"
 
 -- | Get random bytes from the system. Do not over use this function
 -- as it is meant to be used by a PRG. This function reads bytes from