+++ spec/entropyquality.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+stack exec raaz entropy | dieharder -a -g 200