+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Block size moved to type level. Allows better type safety.
 - Use Data.Proxy.Proxy to get rid of some uses of undefined.
 - Support withdrawn for base < 4.7. This allows cleaning up some Nat codes.
+- Linux: getrandom now uses syscall directly, so works even when glibc is old.
+- raaz command expose an entropy command for checking the quality of
+  system entropy source.
 
 ## [0.2.0] - 24 August, 2017